gl2d: Switch to uint16_t for UV coordinates
Using "unsigned int" is very wasteful, as there is never a situation where
the coordinates are bigger than 1024, which fits in 2 bytes.

NOTE: This is a breaking change, but I'm doing it because I don't think
people are actually using this due to the lack of documentation on how
to generate packed textures.

+++ include/gl2d.h
@@ -295,7 +295,7 @@ void glSpriteOnQuad(int x1, int y1, int x2, int y2, int x3, int y3, int x4, int 
 /// @param texture Pointer to the texture data to load.
 /// @return Returns the texture ID on success, -1 on error.
 int glLoadSpriteSet(glImage *sprite, const unsigned int numframes,
-                     const unsigned int *texcoords, GL_TEXTURE_TYPE_ENUM type,
+                     const uint16_t *texcoords, GL_TEXTURE_TYPE_ENUM type,
                      int sizeX, int sizeY, int param, int palette_width,
                      const void *palette, const void *texture);
